+++ Mods/S7_Journal_6b1aee95-bd4c-4190-a919-4aa7096eaa2b/Story/RawFiles/Lua/BootstrapClient.lua
@@ -72,6 +72,7 @@ Ext.RegisterListener("StatsLoaded", function()
 
     local targetTreasureTables = {
         "ST_IngredientsTrader",
+        "ST_KitchenThings",
         "ST_PaperWork"
     }