+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### CHANGED
 
 * _Renamed_ "Journal" to "Notebook".
+
 ## [0.3.0.0] --- 17th December 2020 --- **_Casual rewrite of the entire code_**
 
 ### NEW